Phase 1: Foundation setup for routing refactor
- Created new directory structure: serv/http/, serv/routing/, serv/app/, serv/di/
- Added compatibility layer with re-exports to maintain backward compatibility
- Renamed conflicting modules: app.py -> _app.py, routing.py -> _routing.py
- Updated all imports throughout codebase to use new module names
- New import paths work: serv.http.GetRequest, serv.routing.Router, serv.app.App
- Old import paths still work for backward compatibility
- Basic test infrastructure updated and validated
