+++ planning/auth-system/00-overview.md
@@ -1,0 +1,160 @@
+# Authentication System Architecture Plan
+
+## Overview
+
+This document outlines the design and implementation plan for a comprehensive, extensible authentication and authorization system for the Serv framework. The system is designed to be modular, configurable, and audit-compliant.
+
+## Core Requirements (Initial Scope: Password + Token Auth)
+
+1. **Extensibility**: All components must be pluggable via abstract base classes
+2. **Configuration**: Full configuration through `serv.config.yaml`
+3. **Container Integration**: All providers registered with the app's DI container
+4. **Route Protection**: Policy-based route and router protection
+5. **Comprehensive Auditing**: Mandatory audit logging with compliance enforcement
+6. **Authentication Methods**: Password-based and token-based authentication only
+7. **Testability**: Layered design with clear separation of concerns
+
+## High-Level Architecture
+
+```
+┌─────────────────────────────────────────────────────────────────┐
+│                        Serv Application                         │
+├─────────────────────────────────────────────────────────────────┤
+│  Route Protection Layer (Middleware/Decorators)                │
+├─────────────────────────────────────────────────────────────────┤
+│  Authentication & Authorization Core                           │
+│  ┌─────────────────┐ ┌─────────────────┐ ┌─────────────────┐   │
+│  │   Auth Provider │ │ Policy Engine   │ │ Audit Engine    │   │
+│  └─────────────────┘ └─────────────────┘ └─────────────────┘   │
+├─────────────────────────────────────────────────────────────────┤
+│  Provider Layer                                                │
+│  ┌─────────────┐ ┌─────────────┐ ┌─────────────┐ ┌───────────┐ │
+│  │ Credential  │ │   Session   │ │    User     │ │   Audit   │ │
+│  │  Provider   │ │  Provider   │ │  Provider   │ │ Provider  │ │
+│  └─────────────┘ └─────────────┘ └─────────────┘ └───────────┘ │
+├─────────────────────────────────────────────────────────────────┤
+│  Storage/Backend Layer                                         │
+│  ┌─────────────┐ ┌─────────────┐ ┌─────────────┐ ┌───────────┐ │
+│  │  Database   │ │    Redis    │ │    LDAP     │ │   Files   │ │
+│  └─────────────┘ └─────────────┘ └─────────────┘ └───────────┘ │
+└─────────────────────────────────────────────────────────────────┘
+```
+
+## Implementation Phases
+
+### Phase 1: Foundation & Audit System
+- Base audit enforcement system
+- Abstract provider interfaces
+- Configuration schema
+- Core exceptions and types
+
+### Phase 2: Provider Implementations
+- Credential provider (password + token authentication)
+- Session provider (memory, database)
+- User provider (database, memory)
+- Audit provider (memory, database)
+
+### Phase 3: Auth Core & Policy Engine
+- Authentication orchestration
+- Policy definition and evaluation
+- Permission checking system
+- Route/router protection integration
+
+### Phase 4: Integration & Testing
+- Middleware integration
+- Extension configuration
+- Comprehensive test suite
+- Documentation and examples
+
+## Key Design Principles
+
+1. **Fail Secure**: Default to deny access when in doubt
+2. **Audit Everything**: All auth operations must be audited
+3. **Principle of Least Privilege**: Minimal permissions by default
+4. **Configuration-Driven**: Behavior controlled via config files
+5. **Testable**: Each layer independently testable
+6. **Performance**: Efficient caching and minimal overhead
+
+## File Structure
+
+```
+serv/auth/
+├── __init__.py
+├── types.py              # Core types and enums
+├── exceptions.py         # Auth-specific exceptions
+├── factory.py            # Provider resolution and factory
+├── audit/
+│   ├── __init__.py
+│   ├── enforcement.py    # Audit enforcement system
+│   ├── events.py         # Standard audit events
+│   └── decorators.py     # Audit decorators
+├── providers/
+│   ├── __init__.py
+│   ├── base.py           # Abstract base classes
+│   ├── auth.py           # Auth provider interface
+│   ├── credential.py     # Credential provider interface
+│   ├── session.py        # Session provider interface
+│   ├── user.py           # User provider interface
+│   └── audit.py          # Audit provider interface
+├── core/
+│   ├── __init__.py
+│   ├── auth.py           # Main authentication orchestrator
+│   ├── policy.py         # Policy engine
+│   └── protection.py     # Route protection system
+├── middleware/
+│   ├── __init__.py
+│   ├── auth.py           # Authentication middleware
+│   └── policy.py         # Policy enforcement middleware
+├── config/
+│   ├── __init__.py
+│   ├── schema.py         # Configuration validation
+│   └── loader.py         # Configuration loading
+└── integration/
+    ├── __init__.py
+    ├── extension.py      # Extension system integration
+    └── routing.py        # Routing system integration
+
+serv/bundled/auth/
+├── __init__.py
+├── memory/               # In-memory implementations
+│   ├── __init__.py
+│   ├── credential.py     # MemoryCredentialProvider
+│   ├── session.py        # MemorySessionProvider
+│   ├── user.py           # MemoryUserProvider
+│   └── audit.py          # MemoryAuditProvider
+├── database/             # Database implementations
+│   ├── __init__.py
+│   ├── credential.py     # DatabaseCredentialProvider
+│   ├── session.py        # DatabaseSessionProvider
+│   ├── user.py           # DatabaseUserProvider
+│   ├── audit.py          # DatabaseAuditProvider
+│   └── schema.py         # Database schema definitions
+├── redis/                # Redis implementations
+│   ├── __init__.py
+│   ├── session.py        # RedisSessionProvider
+│   └── audit.py          # RedisAuditProvider
+# Additional providers for future expansion:
+# ├── ldap/                 # LDAP implementations (future)
+# ├── oauth/                # OAuth implementations (future)
+# └── saml/                 # SAML implementations (future)
+├── auth/                 # Main auth implementations
+│   ├── __init__.py
+│   └── standard.py       # StandardAuthProvider
+└── policy/               # Policy implementations
+    ├── __init__.py
+    ├── rbac.py           # RBACPolicyProvider
+    └── abac.py           # ABACPolicyProvider
+```
+
+## Next Steps
+
+1. Review and approve this architectural plan
+2. Define detailed interfaces and contracts
+3. Implement audit enforcement system (most critical)
+4. Build abstract provider interfaces
+5. Create configuration schema
+6. Implement core providers
+7. Build authentication orchestrator
+8. Integrate with routing system
+9. Comprehensive testing
+10. Documentation and examples
