+++ planning/auth-system/README.md
@@ -1,0 +1,187 @@
+# Authentication System Planning
+
+## Overview
+
+This directory contains the comprehensive planning documentation for implementing a production-ready authentication and authorization system for the Serv framework.
+
+**Initial Scope**: Password and token-based authentication with memory and database providers. External authentication providers (OAuth, LDAP, SAML) will be added in future versions.
+
+## Planning Documents
+
+### [00-overview.md](./00-overview.md) - System Architecture Overview
+- High-level architecture and requirements
+- Core components and their relationships
+- Implementation phases and principles
+- File structure and organization
+
+### [01-audit-system.md](./01-audit-system.md) - Audit System Design
+- Comprehensive audit enforcement system
+- Audit event types and data structures
+- Mandatory audit decorators and inheritance enforcement
+- Compliance and security logging requirements
+
+### [02-provider-interfaces.md](./02-provider-interfaces.md) - Provider Interface Design
+- Abstract base classes for all providers
+- Complete method signatures with audit requirements
+- Type definitions and data structures
+- Error handling and configuration patterns
+
+### [03-configuration-schema.md](./03-configuration-schema.md) - Configuration Design
+- Complete `serv.config.yaml` schema for authentication
+- Extension-level `extension.yaml` policy configuration
+- Policy expression language and validation
+- Environment variable integration
+
+### [04-route-protection.md](./04-route-protection.md) - Route Protection System
+- Authentication and authorization middleware
+- Policy engine and route protection mechanisms
+- Configuration-based route protection via extension.yaml
+- Router-level access control
+
+### [05-implementation-plan.md](./05-implementation-plan.md) - Implementation Roadmap
+- Detailed 5-phase implementation plan
+- Task breakdown and dependencies
+- Quality gates and success metrics
+- Risk mitigation strategies
+
+### [06-testing-strategy.md](./06-testing-strategy.md) - Comprehensive Testing Plan
+- Unit, integration, and end-to-end testing strategies
+- Security and penetration testing approaches
+- Performance and load testing requirements
+- Test utilities and mock implementations
+
+## Key Features
+
+### 🔐 Security-First Design
+- **Fail Secure**: Default deny access policy
+- **Comprehensive Auditing**: All security operations audited
+- **Audit Enforcement**: Compile-time and runtime audit validation
+- **Defense in Depth**: Multiple layers of security controls
+
+### 🔧 Extensible Architecture
+- **Provider Pattern**: Pluggable authentication backends
+- **Configuration-Driven**: Full control via YAML configuration
+- **Container Integration**: Native dependency injection support
+- **Core Auth Methods**: Password and token-based authentication
+
+### 🎯 Route Protection
+- **Granular Control**: Route and router-level protection
+- **Policy Engine**: Flexible permission and role-based policies
+- **Middleware Integration**: Seamless framework integration
+- **Performance Optimized**: Caching and efficient evaluation
+
+### 📊 Compliance Ready
+- **Audit Trail**: Complete security event logging
+- **Data Protection**: GDPR and privacy compliance features
+- **Retention Policies**: Configurable data retention
+- **Regulatory Standards**: SOC 2, PCI-DSS support
+
+## Core Components
+
+### Provider Types
+1. **Auth Provider** - Main authentication orchestrator
+2. **Credential Provider** - Credential storage and verification
+3. **Session Provider** - Session management and storage
+4. **User Provider** - User data and role management
+5. **Audit Provider** - Security event logging and monitoring
+
+### Protection Mechanisms
+- **Authentication Middleware** - Request-level authentication
+- **Policy Engine** - Permission evaluation and caching
+- **Configuration-Based Protection** - Route protection via extension.yaml
+- **Router Guards** - Extension-level access control
+
+### Configuration System
+- **Provider Configuration** - Backend service configuration
+- **Policy Configuration** - Access control rules
+- **Security Settings** - Rate limiting, headers, validation
+- **Integration Settings** - External service configuration
+
+## Implementation Approach
+
+### Phase 1: Foundation (Week 1)
+- Core types and audit enforcement system
+- Provider interfaces and base implementations
+- Configuration schema and validation
+
+### Phase 2: Basic Providers (Week 2)
+- Memory and database provider implementations
+- Testing infrastructure and mock providers
+- Integration testing framework
+
+### Phase 3: Configuration & Policy (Week 3)
+- Configuration loading and validation
+- Policy engine and RBAC implementation
+- Permission evaluation system
+
+### Phase 4: Auth Core (Week 3-4)
+- Main authentication provider
+- Session management integration
+- Multi-factor authentication support
+
+### Phase 5: Route Protection (Week 4)
+- Middleware implementation
+- Extension integration system
+- Router protection mechanisms
+
+### Phase 6: Enhanced Features (Week 5)  
+- Enhanced password validation and token management
+- Basic security features (input validation, secure handling)
+- Performance optimization and polish
+
+## Quality Assurance
+
+### Testing Requirements
+- **90%+ Code Coverage** - Comprehensive test coverage
+- **Security Testing** - Penetration testing and vulnerability assessment
+- **Performance Testing** - Load testing and benchmarking
+- **Compliance Testing** - Audit trail and regulatory compliance
+
+### Code Quality
+- **Type Safety** - Complete type hints throughout
+- **Documentation** - Comprehensive API documentation
+- **Security Review** - Multiple security review cycles
+- **Performance Optimization** - Caching and efficient algorithms
+
+## Getting Started
+
+1. **Review Architecture** - Start with [00-overview.md](./00-overview.md)
+2. **Understand Audit System** - Review [01-audit-system.md](./01-audit-system.md) 
+3. **Study Interfaces** - Examine [02-provider-interfaces.md](./02-provider-interfaces.md)
+4. **Check Configuration** - Review [03-configuration-schema.md](./03-configuration-schema.md)
+5. **Follow Implementation Plan** - Use [05-implementation-plan.md](./05-implementation-plan.md)
+
+## Design Principles
+
+1. **Security by Default** - Secure defaults and fail-secure behavior
+2. **Principle of Least Privilege** - Minimal permissions by default
+3. **Defense in Depth** - Multiple security layers
+4. **Audit Everything** - Complete security event logging
+5. **Configuration-Driven** - Behavior controlled via configuration
+6. **Performance Conscious** - Efficient and scalable design
+7. **Standards Compliant** - Industry security standards adherence
+
+## Success Criteria
+
+### Functional Requirements
+- ✅ Complete provider interface implementation
+- ✅ Configuration-driven authentication and authorization
+- ✅ Route and router protection system
+- ✅ Comprehensive audit trail
+- ✅ Multi-provider support (database, memory, external)
+
+### Quality Requirements
+- ✅ 90%+ test coverage
+- ✅ Zero security vulnerabilities
+- ✅ Performance within acceptable limits
+- ✅ Complete documentation
+- ✅ Compliance with security standards
+
+### Integration Requirements
+- ✅ Seamless Serv framework integration
+- ✅ Extension system compatibility
+- ✅ Container-based dependency injection
+- ✅ Middleware and routing integration
+- ✅ Configuration system integration
+
+This authentication system will provide a robust, secure, and extensible foundation for authentication and authorization in Serv applications, meeting enterprise security requirements while maintaining developer productivity and system performance.
