+++ planning/routing-refactor/README.md
@@ -1,0 +1,36 @@
+# Routing Code Refactoring Project
+
+This directory contains planning documents for refactoring the routing code organization in Serv to improve modularity, maintainability, and clarity.
+
+## Overview
+
+The current routing implementation has grown organically and suffers from several organizational issues:
+- Overly large files (`routes.py` at 1,138 lines, `app.py` at 1,112 lines)
+- Mixed concerns (HTTP handling, routing logic, form processing all in one place)
+- Unclear module boundaries
+- Difficult to test and maintain
+
+## Planning Documents
+
+1. **00-current-analysis.md** - Detailed analysis of the current routing code structure
+2. **01-proposed-structure.md** - New module organization and responsibilities
+3. **02-migration-strategy.md** - Step-by-step plan for refactoring without breaking changes
+4. **03-testing-approach.md** - How to ensure we don't break existing functionality
+5. **04-implementation-checklist.md** - Detailed action items and progress tracking
+
+## Goals
+
+- **Modularity**: Break large files into focused, single-responsibility modules
+- **Clarity**: Clear separation between HTTP, routing, and application concerns
+- **Testability**: Smaller modules that are easier to unit test
+- **Maintainability**: Logical organization that's easier to understand and modify
+- **Backward Compatibility**: No breaking changes to public APIs
+
+## Success Criteria
+
+- [ ] No file over 500 lines
+- [ ] Clear module boundaries with single responsibilities
+- [ ] All existing tests pass
+- [ ] New structure is documented
+- [ ] Import paths are clean and logical
+- [ ] Performance is maintained or improved
