+++ serv/_app.py
@@ -36,7 +36,7 @@ from serv.injectors import inject_request_object, inject_websocket_object
 from serv.protocols import AppContextProtocol, EventEmitterProtocol, RouterProtocol
 from serv.requests import Request
 from serv.responses import ResponseBuilder
-from serv.routing import HTTPNotFoundException, Router
+from serv._routing import HTTPNotFoundException, Router
 
 logger = logging.getLogger(__name__)
 
@@ -979,7 +979,7 @@ class App(EventEmitterProtocol, AppContextProtocol):
                 # Create a branch of the container with route settings
                 with container.branch() as route_container:
                     # Add route settings to the container using RouteSettings
-                    from serv.routing import RouteSettings
+                    from serv._routing import RouteSettings
 
                     route_container.add(RouteSettings, RouteSettings(**route_settings))
 
@@ -1080,7 +1080,7 @@ class App(EventEmitterProtocol, AppContextProtocol):
 
                 # Create a branch of the container with route settings and WebSocket instance
                 with container.branch() as route_container:
-                    from serv.routing import RouteSettings
+                    from serv._routing import RouteSettings
 
                     route_container.add(RouteSettings, RouteSettings(**route_settings))
                     route_container.add(WebSocket, websocket)
