+++ serv/di/__init__.py
@@ -1,0 +1,5 @@
+"""Dependency injection layer for Serv - parameter injection, containers, resolvers."""
+
+# Initially empty - will be populated during Phase 5
+
+__all__ = []
