+++ serv/http/__init__.py
@@ -1,0 +1,30 @@
+"""HTTP layer for Serv - request/response handling, forms, validation."""
+
+# Re-export from existing modules for backward compatibility
+try:
+    from serv.routes import GetRequest, PostRequest, PutRequest, DeleteRequest, PatchRequest
+    from serv.responses import (
+        JSONResponse,
+        HTMLResponse, 
+        TextResponse,
+        RedirectResponse,
+        Response,
+        ResponseBuilder,
+    )
+    
+    __all__ = [
+        "GetRequest",
+        "PostRequest", 
+        "PutRequest",
+        "DeleteRequest",
+        "PatchRequest",
+        "JSONResponse",
+        "HTMLResponse",
+        "TextResponse",
+        "RedirectResponse",
+        "Response",
+        "ResponseBuilder",
+    ]
+except ImportError:
+    # If direct imports fail, will be populated later during refactoring
+    __all__ = []
