+++ tests/e2e/helpers.py
@@ -13,7 +13,7 @@ from typing import Any
 
 from httpx import ASGITransport, AsyncClient
 
-from serv.app import App
+from serv._app import App
 from serv.extensions import Extension
