+++ tests/test_app_welcome_extension.py
@@ -5,7 +5,7 @@ import pytest
 import yaml
 from bevy.registries import Registry
 
-from serv.app import App
+from serv._app import App
 from serv.extensions import Extension
 from serv.extensions.loader import ExtensionSpec
