+++ tests/test_e2e_example.py
@@ -12,11 +12,11 @@ from pathlib import Path
 import pytest
 from bevy import Inject, injectable
 
-from serv.app import App
+from serv._app import App
 from serv.extensions import Extension, on
 from serv.extensions.loader import ExtensionSpec
 from serv.responses import ResponseBuilder
-from serv.routing import Router
+from serv._routing import Router
 from tests.e2e_test_helpers import AppBuilder, create_test_client
