+++ tests/test_extension_settings.py
@@ -10,7 +10,7 @@ from bevy.registries import Registry
 from serv.extensions import Extension, on
 from serv.extensions.loader import ExtensionSpec
 from serv.responses import ResponseBuilder
-from serv.routing import Router
+from serv._routing import Router
 from tests.helpers import create_mock_importer
