+++ tests/test_middleware.py
@@ -4,7 +4,7 @@ import pytest
 from bevy import Inject, injectable
 from httpx import AsyncClient
 
-from serv.app import App
+from serv._app import App
 from serv.exceptions import (
     HTTPNotFoundException,  # For testing middleware error handling
 )
