+++ tests/test_route_file_upload_validation.py
@@ -9,10 +9,10 @@ import pytest
 from bevy import Inject, injectable
 from httpx import AsyncClient
 
-from serv.app import App
+from serv._app import App
 from serv.extensions import Extension, on
 from serv.routes import PostRequest, Route, TextResponse, handle
-from serv.routing import Router
+from serv._routing import Router
 from tests.helpers import create_test_extension_spec
